changes to the ingest pipeline corresponding to the datagram format change in the new TOS (ASKAPSDP-1709, ASKAPSDP-1713, ASKAPTOS-3640); also added card information to the log message of the visibility source which is handy for debugging

+++ Code/Components/Services/common/current/cpcommon/VisDatagramADE.h
@@ -86,12 +86,12 @@ namespace askap {
             
             /// @brief channel
             /// @details frequency channel index. 
-            /// Allowed values are from 0 to 215
+            /// Allowed values are from 1 to 216
             /// Part of frequency index
             uint32_t channel;
             
             /// Frequency - sky frequency in MHz (originated in firmware)
-            float freq;
+            double freq;
 
             /// Beam ID
             /// This identified which synthesised beam this datagram corresponds to.
